# docs/api/api_doc.yml
openapi: 3.0.3
info:
  title: re_place API
  version: 0.1.0
  description: API contract generated from current Next.js app/api routes.
servers:
  - url: http://localhost:3000
...
paths:
  /api/articles:
    get:
      tags: [Articles]
      summary: List and search current user's articles
      parameters:
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [articleTime, createTime, topic, title]
            default: createTime
          description: Sort field. Invalid value falls back to createTime.
        - in: query
          name: search
          schema:
            type: string
          description: Full-text search over title/city/topic/content/tags/AI fields.
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
        "401":
          description: Missing or invalid auth session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

    post:
      tags: [Articles]
      summary: Create article and upload images to Firebase Storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleCreateRequest"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleCreateSuccessResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          description: Missing or invalid auth session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  schemas:
...
    ErrorMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    ArticleSortBy:
      type: string
      enum: [articleTime, createTime, topic, title]
    ArticleCreateRequest:
      type: object
      required: [content, imageUrls, useAiFineTune]
      properties:
        title:
          type: string
        content:
          type: string
          description: Max 600 characters.
        imageUrls:
          type: array
          maxItems: 3
          items:
            type: string
            description: Base64 data URL, e.g. data:image/png;base64,...
        useAiFineTune:
          type: boolean
    ArticleCreateSuccessResponse:
      type: object
      required: [message, accepted, entry]
      properties:
        message:
          type: string
        accepted:
          type: boolean
        entry:
          $ref: "#/components/schemas/HistoryEntry"
    HistoryEntryAiExtracted:
      type: object
      required: [summary, keywords, places]
      properties:
        summary:
          type: string
        keywords:
          type: array
          items:
            type: string
        places:
          type: array
          items:
            type: string
    HistoryEntry:
      type: object
      required:
        [id, title, city, topic, tags, content, imageUrls, aiExtracted, articleTime, createTime]
      properties:
        id:
          type: string
        title:
          type: string
        city:
          type: string
        topic:
          type: string
        tags:
          type: array
          items:
            type: string
        content:
          type: string
        imageUrls:
          type: array
          items:
            type: string
            format: uri
        aiExtracted:
          $ref: "#/components/schemas/HistoryEntryAiExtracted"
        articleTime:
          type: string
          format: date-time
        createTime:
          type: string
          format: date-time
    ArticleListResponse:
      type: object
      required: [sortBy, search, entries]
      properties:
        sortBy:
          $ref: "#/components/schemas/ArticleSortBy"
        search:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
...